+++ eddrit/reddit/content_parser/media.py
@@ -72,14 +72,22 @@ def get_post_gallery_content(
     try:
         # Get image urls in order
         gallery_data = api_post_data["gallery_data"].get("items")
-        sorted_image_ids = [item["media_id"] for item in gallery_data]
+        sorted_image_ids_and_captions = [
+            (item["media_id"], item.get("caption", "")) for item in gallery_data
+        ]
 
         media_metadata = api_post_data["media_metadata"]
-        sorted_image_metadata = [media_metadata[id] for id in sorted_image_ids]
+        sorted_image_metadata_and_caption = [
+            (media_metadata[id], caption)
+            for id, caption in sorted_image_ids_and_captions
+        ]
 
-        pictures = [_get_gallery_picture(item) for item in sorted_image_metadata]
+        pictures = [
+            _get_gallery_picture(item) for item, _ in sorted_image_metadata_and_caption
+        ]
+        captions = [caption for _, caption in sorted_image_metadata_and_caption]
 
-        return models.GalleryPostContent(pictures=pictures)
+        return models.GalleryPostContent(pictures=pictures, captions=captions)
     except Exception:
         return models.LinkPostContent()